+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### next [????.??.??]
 * Add `Data.Deriving.Via`, which allows emulating the behavior of the
   `GeneralizedNewtypeDeriving` and `DerivingVia` extensions.
+* Test suite fixes for GHC 8.4.
 
 ## 0.4 [2017.12.07]
 * Incorporate changes from the `EmptyDataDeriving` proposal (which is in GHC